Add customization of analog eyes blinking threshold
An analog fist can have an animation with eyes closed and eyes open.
Add an option in the Compiler to customize the threshold after which eyes are closed.

The option was added in the Compiler for simplicity, as if it were on Activities, it would require special handling for deduplication in the optimization phase for the blinking combiner.

Closes #5